refactor(nuxt-app): :recycle: move examples to separate layers
