+++ examples/nuxt-app/layers/example-components/components/global/TableExtraContentComponentExample.vue
