+++ examples/nuxt-app/layers/example-components/components/global/TableExtraContentItemComponentExample.vue
