+++ examples/nuxt-app/layers/example-components/components/global/TideSearchBarComponentExample.vue
