+++ examples/nuxt-app/layers/example-components/components/global/TideSearchBelowFilter.vue
