+++ examples/nuxt-app/layers/example-components/components/global/TideSearchEmpty.vue
