+++ examples/nuxt-app/layers/example-components/components/global/TideSearchResultExampleSkeleton.vue
