+++ examples/nuxt-app/layers/fixture-api/pages/_fixture/tide-site.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script setup lang="ts">
-import page from '../test/fixtures/case/accordions-inpage-nav.json'
+import page from '../../test/fixtures/page/accordions-inpage-nav.json'
 import { useTideSite } from '#imports'
 
 const site = await useTideSite()
