+++ examples/nuxt-app/layers/map-features/components/global/VSBAProjectAreaLayer.vue
