+++ examples/nuxt-app/layers/map-features/components/global/VsbaMapPopupContent.vue
