+++ examples/nuxt-app/layers/map-features/components/global/test/TestMapShapeLayer.vue
