+++ examples/nuxt-app/layers/map-features/components/global/test/TestMapSidepanelItem.vue
