+++ examples/nuxt-app/layers/ripple-ui-forms-ext/README.md
@@ -1,0 +1,6 @@
+## About this layer
+
+This layer includes an example for extending `ripple-ui-forms`.
+
+- Drupal webform control is `item`
+- Backend is fixtured for page and site for [test page](http://localhost:3000/_test/webform)
