+++ examples/nuxt-app/layers/ripple-ui-forms-ext/pages/_test/webform.vue
@@ -6,7 +6,5 @@
 
 <script setup lang="ts">
 import page from '../../__test__/fixtures/webform.json'
-import { useTideSite } from '#imports'
-
-const site = await useTideSite()
+import site from '../../__test__/fixtures/site.json'
 </script>
